--- a/python_modules/dagster/dagster/_core/storage/alembic/versions/043_add_run_tags_run_id_index.py
+++ b/python_modules/dagster/dagster/_core/storage/alembic/versions/043_add_run_tags_run_id_index.py
@@ -27,9 +27,7 @@ def upgrade():
             ["run_id", "id"],
             unique=False,
             postgresql_concurrently=True,
-            mysql_length={
-                "run_id": 255
-            },
+            mysql_length={"run_id": 255},
         )