@prha 2024-07-09
Yeah, we've known for a little while now that we need to join on `runs.id`... We've punted on it because of the complexity involved in managing the opt-in migrations.  But I think it is probably worth doing because of how the performance degrades over time.  

I think this PR makes sense, but I might spend some time this week scoping out what the foreign-key change would look like (involving both a schema and data migration).

This change probably makes sense only if they've created the index.  That tells me that this should probably be two separate PRs, one which creates the schema change with the additional index, and another change that switches the join in the query, which we'd want to coordinate with a major release (e.g. `1.8.0`).