@egordm 2024-07-11
Sounds like a good plan. I will split this PR into two, to start with that. 

Combining the new query with a change in foreign key in the future would be the best.