@egordm 2024-07-17
Hey, I have changed this PR to only add an index and a migration. As described I have also created a second PR #23050 which changes the query itself. Let me know if any other changes need to be made.